minor fixes for 64bit builds; installer for agent 0.2.27 x64

+++ src/agent/core/policy.cpp
@@ -40,7 +40,7 @@ static DWORD DeployConfig(DWORD session, CSCPMessage *msg)
 
 	msg->GetVariableStr(VID_CONFIG_FILE_NAME, name, MAX_PATH);
 	DebugPrintf(session, _T("DeployConfig(): original file name is %s"), name);
-	for(i = _tcslen(name) - 1; i >= 0; i--)
+	for(i = (int)_tcslen(name) - 1; i >= 0; i--)
 		if ((name[i] == '/') || (name[i] == '\\'))
 		{
 			break;
@@ -119,7 +119,7 @@ static DWORD RemoveConfig(DWORD session, CSCPMessage *msg)
 
 	msg->GetVariableStr(VID_CONFIG_FILE_NAME, name, MAX_PATH);
 	DebugPrintf(session, _T("RemoveConfig(): original file name is %s"), name);
-	for(i = _tcslen(name) - 1; i >= 0; i--)
+	for(i = (int)_tcslen(name) - 1; i >= 0; i--)
 		if ((name[i] == '/') || (name[i] == '\\'))
 		{
 			break;
